requirements.txt
dj-database-url==0.5.0
django-daterange-filter==1.3.0
django-extensions==2.2.9
django-localflavor==3.0.1
django-widget-tweaks==1.4.8
Django==2.2.12
Pillow==7.1.1
python-decouple==3.3
python-stdnum==1.13
selenium==3.141.0